update_filter_and_sort
